Move checkpoint bench tests to benchutil package
Eliminates the checkpoint_test external test package by moving
bench_test.go and parse_tree_bench_test.go into benchutil, where
the circular dependency is already resolved.

Entire-Checkpoint: 6783ce272531
