+++ cmd/entire/cli/benchutil/parse_tree_bench_test.go
@@ -1,4 +1,4 @@
-package checkpoint_test
+package benchutil
 
 import (
 	"fmt"
